minot refactor
